KT-09: Importing CSV file as questions with answers

+++ src/main/java/com/wojcka/exammanager/models/Question.java
@@ -36,7 +36,7 @@ public class Question {
 
     @Valid
     @NotNull
-    public Long points;
+    public Integer points;
 
     @Column(nullable = false)
     public Boolean archived = false;
